conversion from int3 to Long

+++ src/io/github/zhtmf/converters/IntegerConverter.java
@@ -62,7 +62,7 @@ public class IntegerConverter implements Converter<Integer> {
             return (int)StreamUtils.readInt(is, ctx.signed, ctx.bigEndian);
         }
         case INT3:{
-            return (int)StreamUtils.readInt3(is, ctx.signed, ctx.bigEndian);
+            return StreamUtils.readInt3(is, ctx.signed, ctx.bigEndian);
         }
         case CHAR:{
             return (int)Utils.deserializeAsCHAR(is, ctx, self, DataType.INT);
